login function

+++ stardew_villager_manager/src/main/java/com/revature/controllers/VillagerController.java
@@ -15,43 +15,63 @@ public class VillagerController {
 	
 	
 	public Handler getVillagersHandler = (ctx) -> {
-		ArrayList<Villagers> villagers = vDAO.getVillagers();
-		Gson gson = new Gson();
-		String JSONvillagers = gson.toJson(villagers);
-		ctx.result(JSONvillagers);
-		ctx.status(200);
+		if(AuthController.ses != null) {
+			ArrayList<Villagers> villagers = vDAO.getVillagers();
+			Gson gson = new Gson();
+			String JSONvillagers = gson.toJson(villagers);
+			ctx.result(JSONvillagers);
+			ctx.status(200);
+		} else {
+			ctx.result("You are not logged in");
+			ctx.status(401);
+		}
 	};
 	
 	public Handler deleteVillagerHandler = (ctx) -> {
-		int villagerToDelete = Integer.valueOf(ctx.pathParam("villager_id"));
-		vDAO.deleteVillager(villagerToDelete);
+		if(AuthController.ses != null) {
+			int villagerToDelete = Integer.valueOf(ctx.pathParam("villager_id"));
+			vDAO.deleteVillager(villagerToDelete);
 		
-		ctx.result("A villager has been removed from Stardew Valley.");
-		ctx.status(200);
+			ctx.result("A villager has been removed from Stardew Valley.");
+			ctx.status(200);
+		} else {
+			ctx.result("You are not logged in");
+			ctx.status(401);
+		}
 		
 	};
 	
 	public Handler updateVillagerHandler = (ctx) -> {
-		String oldName = ctx.pathParam("old_name");
-		String newName = ctx.body();
+		if(AuthController.ses != null) {
+			String oldName = ctx.pathParam("old_name");
+			String newName = ctx.body();
 		
 		
-		if(vDAO.updateVillagerName(newName, oldName)) {
-			ctx.status(202);
+			if(vDAO.updateVillagerName(newName, oldName)) {
+				ctx.status(202);
+			} else {
+				ctx.status(406);
+			}
 		} else {
-			ctx.status(406);
+			ctx.result("You are not logged in");
+			ctx.status(401);
 		}
 	};
 
 	public Handler insertVillagerHandler = (ctx) -> {
-		String body = ctx.body();
-		Gson gson = new Gson();
-		Villagers newVillager = gson.fromJson(body, Villagers.class);
+		if(AuthController.ses != null) {
+			String body = ctx.body();
+			Gson gson = new Gson();
+			Villagers newVillager = gson.fromJson(body, Villagers.class);
 		
-		if(vDAO.insertVillager(newVillager)) {
-			ctx.status(202);
+			if(vDAO.insertVillager(newVillager)) {
+				ctx.status(202);
+			} else {
+				ctx.status(406);
+			}
 		} else {
-			ctx.status(406);
+			ctx.result("You are not logged in");
+			ctx.status(401);
 		}
 	};
 }
